requirements.txt
torch
torchvision
scikit-learn
numpy
pandas
pillow
tqdm
